Exercises 4.15.-4.23.

+++ part4/bloglist/controllers/blogs.js
@@ -1,9 +1,10 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const middleware = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
-
-  const blogs = await Blog.find({})
+  const blogs = await Blog.
+    find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)
 })
 
@@ -16,23 +17,34 @@ blogsRouter.get('/:id', async (request, response) => {
   }
 })
 
-blogsRouter.post('/', async (request, response) => {
+blogsRouter.post('/', middleware.userExtractor, async (request, response) => {
   const body = request.body
+  const user = request.user
 
   const blog = new Blog({
     title: body.title,
     author: body.author,
     url: body.url,
     likes: body.likes || 0,
+    user: user._id
   })
 
-
   const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
   response.status(201).json(savedBlog)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
+blogsRouter.delete('/:id',  middleware.userExtractor, async (request, response) => {
+
+  const user = request.user
+  const blog = await Blog.findById(request.params.id)
+  if ( blog.user.toString() === user._id.toString() ){
+    await Blog.findByIdAndRemove(request.params.id)
+  } else {
+    return response.status(401).json({ error: 'a blog can be deleted only by the user who added the blog' })
+  }
+
   response.status(204).end()
 })
